Move wechat into payment-methods directory
