+++ client/my-sites/checkout/composite-checkout/payment-methods/wechat.js
